交互基本完成

+++ CeramicCraftSimulation/stdafx.h
@@ -59,4 +59,4 @@
 #include <gl/GL.h>
 #include <gl/GLU.h>
 
-extern  int step;
+extern  float step;
